move translator to own file
